requirements_efk.txt
aiohttp==3.8.3
aiosignal==1.3.1
allennlp==2.10.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.12.0
bleach==6.0.0
blis==0.7.9
boto3==1.26.66
botocore==1.29.66
cached-path==1.1.6
cachetools==4.2.4
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
comm==0.1.3
commonmark==0.9.1
confection==0.0.4
cymem==2.0.7
datasets==2.9.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
exceptiongroup==1.1.0
executing==1.2.0
fairscale==0.4.6
fastjsonschema==2.16.3
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==1.34.0
google-auth==2.16.3
google-cloud-core==1.7.3
google-cloud-storage==2.7.0
google-cloud-translate==2.0.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.9.0
h2==3.2.0
h5py==3.8.0
higher==0.2.1
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.0
hydra-core==1.3.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.1.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jsonnet==0.19.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.1.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
langcodes==3.3.0
lmdb==1.4.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mistune==2.0.5
more-itertools==9.0.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
openai==0.27.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
tabulate==0.9.0
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.17.1
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.14
wandb==0.13.10
wasabi==1.1.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.6
xxhash==3.2.0
yarl==1.8.2
zipp==3.12.1
